Update the JFace.action tests to Junit4
Instead of inheriting TestCase these tests use now a JUnit rule for
setup and teardown.
